+++ src/otx/algorithms/detection/adapters/openvino/task.py
@@ -164,7 +164,7 @@ class BaseInferencerWithConverter(BaseInferencer):
             else:
                 features = (
                     copy.deepcopy(prediction["feature_vector"].reshape(-1)),
-                    copy.deepcopy(prediction["saliency_map"][0]),
+                    self.get_saliency_map(prediction, preprocessing_meta),
                 )
 
             result_handler(id, processed_prediciton, features)
@@ -434,10 +434,12 @@ class OpenVINODetectionTask(IDeploymentTask, IInferenceTask, IEvaluationTask, IO
         if self.model is None:
             raise RuntimeError("load_inferencer failed, model is None")
         _hparams = copy.deepcopy(self.hparams)
-        self.confidence_threshold = float(
-            np.frombuffer(self.model.get_data("confidence_threshold"), dtype=np.float32)[0]
-        )
-        _hparams.postprocessing.confidence_threshold = self.confidence_threshold
+        if _hparams.postprocessing.result_based_confidence_threshold:
+            self.confidence_threshold = float(
+                np.frombuffer(self.model.get_data("confidence_threshold"), dtype=np.float32)[0]
+            )
+            _hparams.postprocessing.confidence_threshold = self.confidence_threshold
+        logger.info(f"Confidence Threshold: {_hparams.postprocessing.confidence_threshold}")
         _hparams.postprocessing.use_ellipse_shapes = self.config.postprocessing.use_ellipse_shapes
         async_requests_num = get_default_async_reqs_num()
         args = [