+++ src/otx/cli/tools/optimize.py
@@ -32,6 +32,7 @@ from otx.cli.utils.importing import get_impl_class
 from otx.cli.utils.io import read_model, save_model_data
 from otx.cli.utils.parser import (
     add_hyper_parameters_sub_parser,
+    get_override_param,
     get_parser_and_hprams_data,
 )
 from otx.core.data.adapter import get_dataset_adapter
@@ -70,7 +71,7 @@ def get_args():
     )
 
     add_hyper_parameters_sub_parser(parser, hyper_parameters)
-    override_param = [f"params.{param[2:].split('=')[0]}" for param in params if param.startswith("--")]
+    override_param = get_override_param(params)
 
     return parser.parse_args(), override_param