+++ tests/e2e/cli/anomaly/test_anomaly_classification.py
@@ -58,7 +58,7 @@ class TestToolsAnomalyClassification:
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)
     def test_otx_train(self, template, tmp_dir_path):
-        otx_train_testing(template, tmp_dir_path, otx_dir, args, deterministic=False)
+        otx_train_testing(template, tmp_dir_path, otx_dir, args, deterministic=True)
 
     @e2e_pytest_component
     @pytest.mark.parametrize("template", templates, ids=templates_ids)